+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for more [Create] recipes
+
 ### Changed
-- Seperated the mappers for [Chipped]
+- Separated the mappers for [Chipped]
 
 ## [5.2.4]
 ### Added